+++ CHANGES.rst
@@ -5,8 +5,8 @@ Changelog
 0.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Fix TextLine field.
+  [cekk]
 
 0.1.0 (2021-02-09)
 ------------------